make request token validation configurable
